[shell] Add a dbus call to activate the "main" application launcher
Summary:
ShellCorona provides a new DBus method "activateLauncherMenu". If that
is invoked it will look through all panels for an applet which has
X-Plasma-Provides with org.kde.plasma.launchermenu. If that also has
a global shortuct it gets activated and no further applet is searched.

The idea behind this DBus call is to provide a way to KWin to activate
the main appliction launcher when the Meta key is pressed.

Reviewers: #plasma, hein, mart

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D2356