Nicer serialization of configuration options
Summary:
The previous approach of having an array of object that
hold information about the current config group and the config keys
and values was not really scalable nor intuitive.

Now, the config is an object that has the group names as keys,
and each group has its own configuration (key, value) pairs.

The config groups are identified as a unix-like path

Reviewers: #plasma, mart

Reviewed By: mart

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D2516