Fix clock in vertical panel
Summary:
The clock in the vertical panel can't fit, it won't word wrap as the
clock is only one word long - so it needs to have a horizontalFit size
mode (without a minimum size) so that the text always fits without being
clipped.

This brings the vertical state in line with the horizontalSmall state
which also sets a minimumPointSize of 0.

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D2644