add osd slot for screen layout changes
Summary: This patch adds an OSD to the dbus object so we can hook up layout changes of the XF86Display button. The method isn't really specific to screen osds, but still called like that (it could be given a generic name, since it's basically just a message and an icon to be specified. This current solution is in line with the rest of the code, though.

Test Plan: Test from kscreen's kded, icon and text show up. Didn't test with multiscreen yet (I think this osd should show up on both screens, will do that once off the train).

Reviewers: #plasma, mart

Reviewed By: mart

Subscribers: broulik, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D2711