[Image Wallpaper] Avoid cross-fading which reveals black background behind images
When simultaneously fading two items (one from fully opaque to transparent and the other
one vice-versa) you can briefly see the black background through the images leading to an ugly flicker.

By only doing this when the old image could be seen (when the new one has a smaller
size) we avoid this in the common case of full-size wallpapers.

Ideally, crossfading would be done using a ShaderEffect.

Differential Revision: https://phabricator.kde.org/D2484