[MPRIS Dataengine] Use UTF-8 for URLs instead of Latin1
While investingating properties of the MPRIS dataengine, I have noitced
that the :xsesam:url property in the metadata was encoded incorrectly. I
have several folders containing music with non-Latin characters, and
they showed up as ????.

After a quick chat with kbroulik on IRC, I decided to make this patch,
which uses UTF-8 instead.

Differential Revision: https://phabricator.kde.org/D2732