Don't constrain width of Text field on translated text
Summary:
This caused it to be cropped on some translations.

This patch only sets the minimum size to force it to line up with the form layout,
but still allow the text to be larger if needed.

Also tidy up other use of directly setting width/anchors in a layout.

BUG:368487

Test Plan:
Checked English layout matched up
Changed text to "asdfasdfadsfasdfljasdfjadlf" and checked that still fit

Reviewers: #plasma, mart

Reviewed By: mart

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D2729