Image wallpaper: Fix verification of a background image file
BackgroundFinder::suffixes(), despite its name, actually returns a list
of glob patterns.  The extension of the image file is checked against
this list to ensure that it is supported, which needs to take
account of that.  Do the check in a new function
BackgroundFinder::isAcceptableSuffix(const QString &suffix).

Reviewers: davidedmundson, #plasma
Reviewed By: davidedmundson, #plasma
Subscribers: plasma-devel
Tags: #plasma
Differential Revision: https://phabricator.kde.org/D2722