Renamed userPrompt to userPromptComponent and add userListComponent id
Summary:
Renamed userPrompt to userPromptComponent and add an id
(userListComponent) to the initialItem Login component.

Disable user list when there are too many users in the system

In sddm 0.14, there's a userModel.disableAvatarsThreshold property
that defines how many users are "too many" in the system so loading of
avatars should be disabled. This commit honors that property, so if
there are more than disableAvatarsThreshold users, then the user list
is disabled and a prompt is shown to ask for a username and password.
If the property is not available (for example, because sddm 0.13 is
being used) then it works as usual.

This threshold is specially useful for large enterprises with NFS
automounted home directories. In such environments, using sddm while
avatars are enabled results in all remote home directories being
mounted in the system before one user is able to log in. This has been
reported to take hours in cases with thousands of users.

By default the threshold has a value of 7, but it can be modified
in sddm.conf, under the Theme section, in the
UsersThresholdToDisableAvatars variable.

Reviewers: #plasma, davidedmundson

Reviewed By: #plasma, davidedmundson

Subscribers: davidedmundson, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D2874