Don't load QPT in kcminit
Summary:
KCMInit is loaded early in startkde, and it blocks all other processes
from starting until it is complete.

This can be up to 1.1 seconds.
(http://imgh.us/bootchart-20160928-1316.svg)

A sizable amount of time (46% of CPU) is spent loading our QPT. This
isn't very useful given none of the kcm_init scripts have any UI and
it's not something they should have. It's only a QGuiApplication so that
they have an X connection.

Test Plan: Checked it ran without errors

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D2882