[shell] Fix non-interactive panelview on non-primary screen
Summary:
If we don't pass the current screen to QCursor::pos, it will return
position relative to primary screen, which obviously won't be in panel
geometry.

Instead pass it currentScreen, this way we don't filter out the mouse
events for panel.

BUG: 368790
FIXED-IN: 5.8.1

Test Plan:
added debug statements and verified that I get correct screen position
and as well, panel is now interactive on secondary screen.

Reviewers: mart, graesslin, #plasma, sebas

Reviewed By: graesslin, #plasma, sebas

Subscribers: sebas, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D2960