No-op SeachRunner if indexing is disabled
Summary: We won't have any results anyway.

Test Plan:
Disabled indexing and put a qDebug in the if branch

I'm not super convinced it's a good idea myself. It means parsing a config file every match
and it can't be cached without missing updates if the user enables/disables at runtime.

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D2995