sddm-theme: support configurable solid colour and image backgrounds
This patch reworks the Background component in the SDDM theme to
support distinct "image" and "color" modes, with a background image
being displayed in the former mode and a solid colour in the latter.

It further adds the background colour and the background type to
the configuration file.

Fixes D2983