Lockscreen: New session only on switchSessionPage
Instead of showing the "New Session" item on the main lockscreen and
the subpage switchSessionPage, we only show it on the second one.

The lockscreen's functionality is in the first place to lock the
screen and therefore it shouldn't largely advertise the possibility
of creating a new session.

That's why we move the creation of a new session completely to the
subpage switchSessionPage, where it fits better to the other
possibility of switching to an already existing one.

Reviewers: davidedmundson

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D3034