Removes preferred definition so a new wallpaper is selected.
Summary:
Removes preferred definition so a new wallpaper is selected.
Also, just in case, check for division by 0.

When the screen resolution changes, a new wallpaper has to be chosen
but this is not happening unless we remove the preferred definition.

This depends on https://git.reviewboard.kde.org/r/129118/ to be accepted.

Test Plan:
I ran plasmashell with the default opensuse wallpaper (a green bulb)
while changing the screen resolution between 1024x768 and 1280x720.

With this and the kpackage patch above, a different background image is
selected correctly with the correct aspect ratio.

Reviewers: #plasma, mart

Reviewed By: mart

Subscribers: mart, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D2981