Support drags from Widget Explorer onto System Tray
Summary:
Some users complain that applets do not behave like applets in the
system tray, this is because they're not adding them in the system tray.

This makes it a bit more intuitive to add applets to the system tray by
allowing drag and drop of applets with NotificationArea=true into the
system tray.

CCBUG: 358283

Test Plan:
Dragged digital clock over system tray
system tray moved and the clock was added to the side

Dragged kate sessions over system tray
on drop, kate sessions applet was added to the tray

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D3212