Disable starting kactivitymanagerd throughout launchertasksmodeltest
Summary:
Currently kactivitymanagerd is loaded halfway through this test, which
has the net result of just breaking.

The test doesn't need it and still passes without it.

Test Plan:
Ran the unit test that's been failing on CI for ages.
It used to fail locally (if run in a new session, see my email on
plasma-devel)
now it passes.

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D3229