Close the system tray on escape - system tray
Summary:
A lot of applets have independent handling making escape close the
expanded representation. However not all of them do.

This leads to both duplicated code and inconsistencies, so we should
tackle this at the source.

If an applet was already using escape internally it will be accepting
the event, and so this code will not cause any regressions.

This patch also removes the stackview forwarding events to the active
child as this doesn't make sense - the active child will already have
focus and get the key event first, so this just created a recursive
chain (which Qt was thankfully fixing for us). Removing it is needed for
this to work.

BUG: 362657

Test Plan:
Pressed escape on various applets (e.g plasma-pa)
Pressed escape on the expander applet list view

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D3257