[Notifications] Show interactive thumbnails on notifications
When a notification contains URLs, they will be shown as a large preview (in case of one
thumbnail) or multiple in a strip (similar to what Klipper does).

The previews can be clicked to open the file and can be dragged anywhere (e.g. to a
webbrowser or chat window).

Differential Revision: https://phabricator.kde.org/D3539