Option to show percentage charge in the icon
Summary:
add an option to show a text label with the percentage value
in the battery icon. text will be shown bottom-right
of the icon with a background for readability, similar to
unread count for IM or email apps.
code was already there but effectively dead code

Test Plan: tried both with and without label

Reviewers: broulik, #plasma

Reviewed By: broulik, #plasma

Subscribers: broulik, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D3603