Refresh the ScreenGeometry data role whenever a screen is added or changes its geometry.
Summary:
As this happens the screen filtering higher up in the proxy chain
culimating in TasksModel needs to be re-executed, as it compares
ScreenGeometry against a property value set on the model. Without
this update, tasks may disappear from views using TasksModel.

Mapping the trigger conditions to affected rows is deliberately
maximally coarse; see implementation comment in code.

BUG:373075

Reviewers: #plasma, davidedmundson

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D3628