Listen for broadcast notifications on system bus
This will listen for a signal on the system DBus to show a notification to all
users on the current machine. This allows prettier and more versatile
notifications than using wall/write.

This is disabled by default but can be enabled in e.g. ~/.config/plasmashellrc
per user or /etc/xdg/plasmashellrc globally:

    [Notifications]
    ListenForBroadcasts=true

Differential Revision: https://phabricator.kde.org/D3606