Fix regression in last commit, don't show battery life as 100 when it's 0
There was code added (approved by me) to make sure we always cast
battery icon percent to an integer if it's undefined. It had the
unintended side effect that a battery life of 0 would also be 'cast' to
100.

When this is undefined we don't show the icon anyway, so
it doesn't really matter what value we fall back to.