[MPRIS Dataengine] Export process id of player instance as separate data entry
The MPRIS specification recommends to media players to distinguish instances of
them by appending its process id. For example this is supported by the Dragon
Player. The patch makes it possible for connected clients of MPRIS Dataengine
to query this directly and by that distinguish multiple instances of a player.

While clients were able to do this earlier already by checking the source name
and doing the same string operations as here, it's more convenient with this
patch, since they then don't have to do it on their own anymore but just need
to query the data.

Reviewers: #Plasma, davidedmundson

Tags: #Plasma

Differential Revision: https://phabricator.kde.org/D3667