introduce new wrapper to disqualify services
disqualify() checks whether the service ought to be disqualified (i.e.
not processed) and then marks it seen. this allows us to boilerplate
call disqualify(service) whenever iterating services to decide whether
the service in question should be skipped.
also means we don't have to call noDisplay once a service is disqualified

reviewed by kbroulik on irc