[Applicationjobs Data Engine] Expose job "destUrl"
This will eventually allow the notification applet to show a better finished notification.
Currently, it just uses whatever was set as second info label last.
While this works for when copying a single file, it's utterly wrong when copying folders
as this will be a random file that was copied last.

Differential Revision: https://phabricator.kde.org/D3819