Clean up KAppMenu
Massively cleanup KAppMenu by removing outdated and unused code.
The global menu has been removed as it will eventually be provided by a plasmoid (famous last words).

The biggest change, however, is that instead of relying on Windows IDs (which won't work on Wayland)
we always speak of DBus service name and object path which will be added as properties to the respective
windows. To support 3rd party applications, we also set the property in the RegisterWindow function.

This way KWin doesn't need any fallback code and can just rely on those properties.

Differential Revision: https://phabricator.kde.org/D3086