[Image Wallpaper] Simplify width/height management
Directly expose "targetSize" as property. The wallpaper is already initialized with
the correct size for some time now but the fact that we had two properties width and height
in there had it still do unneccessary work when it emitted "sizeChanged(2560, 0)" and
then the final "sizeChanged(2560, 1440)" afterwards.

Also, I didn't really see a point in storing the wallpaper size in its config,
it's overwritten by the wallpaper anyway and only used in the wallpaper config dialog
for the aspect ratio. We can just ask the view directly then.

BUG: 374689
FIXED-IN: 5.9.0

Differential Revision: https://phabricator.kde.org/D4001