panel alignment/visibility not resolution dependent
Summary:
panel size and position in pixel is dependent from screen resolution,
this saved in different config groups, as it should.
but two properies of panelview shouldn't change between resolution,
that are visibility and alignment, otherwise the panel changes
behavior in an unpredictable way when resolution changes.

BUG:372064

Test Plan:
started plasmashell with old config, values have been correctly
migrated to the parent group that's common for all resolutions

Reviewers: davidedmundson, #plasma

Reviewed By: davidedmundson, #plasma

Subscribers: davidedmundson, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D4087