Keep AppletMenu as small as possible
Summary:
In retrospect there's no need to fill the width, if people want that
effect they can add a spacer; whereas you can't do the opposite.

Also set the size to 0 when it's empty. This mean using status as
normally PanelContainment ensures plasmoid's are a minimum of a square.

BUG: 374996

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D4115