Fix mapToGlobal
Summary:
When recreating QQuickItem::mapToGlobal for < Qt 5.7 we are currently
using the window's global position and the item's position relevant to
it's parent. We need to use the item's position relative to the scene.

Test Plan: Had applet on right hand side of panel, menus are now in the correct place.

Reviewers: #plasma, broulik

Reviewed By: #plasma, broulik

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D4117