Enable Edit Action in plasmoid mode
Summary:
EditAction is a keyboard activated shortcut that shows a dialog, there's
no reason for it to be available in standalone mode only, especially given the
option to enable the shortcut is available from the plasmoid.

BUG: 374961

Test Plan:
Set shortcut.
Pressed it.
Got a dialog with my clipboard contents

Standalone mode should be unchanged as it was showing it already.

Reviewers: #plasma, mart

Reviewed By: mart

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D4114