Fix clang address sanitiser error in SNI memcpy.
Summary:
If an image is created with no size, QImage::data() points to nullptr.
We then pass this as an argument to memcpy. It doesn't really matter
because we're copying 0 bytes into it, but it's still thrown as an
error.

plasma-
workspace/dataengines/statusnotifieritem/statusnotifieritemsource.cpp:401:12:
runtime error: null pointer passed as argument 1, which is declared to
never be null

Test Plan: Ran it again. No error

Reviewers: #plasma, apol

Reviewed By: apol

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D4196