Remove unused feature "favorite" from PlasmaAppletItemModel
Summary:
There was no role exposed for isFavorite and it has no other impact.
Removing it not only cleans up the code but also saves us parsing a
config file for no reason.

I don't think it's worth putting in the UI; no-one is adding so many
plasmoids that they would go out of their way to mark one as a favourite
for quicker future access.

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D4217